requirements.txt
smolagents
fastapi
uvicorn
python-dotenv
aiofiles
aiohttp
beautifulsoup4
pandas
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
python-multipart
pydantic
email-validator==2.1.0.post1
Jinja2
litellm
schedule
duckduckgo_search
